backend/requirements.txt
fastapi
uvicorn
pytest
yfinance
requests
python-dotenv
pandas
